Move the android data dir retrieving code out of DCHECK.
In CL http://codereview.chromium.org/10035034 we added this code to initialize
resource pak files with the android data dir. By mistake the call to retrieve
the data dir was put within a DCHECK so it doesn't get called in a release or
official build.

BUG=none
TEST=android content shell in an official build loads the content_shell.pak file properly.


Review URL: https://chromiumcodereview.appspot.com/10703109