Don't generate password for GAIA forms.
Note that I haven't written a specific test for this change. It seems impossible to create a form in the password_generatiion_manager_browsertest.cc that matches the GAIA's realm without changing LoadHTML() and the code to create the password form in WebPasswordFormData.cpp (And I don't think we should change them merely for this change). I have verified that this change takes effect in Chrome.

BUG=131322
TEST=PasswordGenerationManagerTest


Review URL: https://chromiumcodereview.appspot.com/10771003