[android] Upstream / sync most of build/android and build/android/pylib.
These files have diverged overtime. We need to get them in sync in
preparation for the instrumentation tests.
The patches downstream have been entangled, so this is a bit big and
contains a series of otherwise unrelated patches.
However, it's probably safer to do this way (as it's guaranteed to be similar to
downstream), than trying to split it in multiple patches.

BUG=
TEST=try -b android_test


Review URL: https://chromiumcodereview.appspot.com/10689132