Revert 145892 - Send the user's Windows profile type up in crash reports.
I suspect that some of the version shear crashes happen to users with non-local profiles.  In this case, it's easy to imagine that the version info in the registry is out of sync with that on the filesystem.

BUG=78585
TEST=none

Review URL: https://chromiumcodereview.appspot.com/10740003

TBR=grt@chromium.org
Review URL: https://chromiumcodereview.appspot.com/10704132