Fix race in HttpNetworkTransactionSpdy2Test.BasicAuthSpdyProxy and re-enable it on TSAN/Dr. Memory.  Apply same fix to the corresponding Spdy 3 test.
Issue was that the tests relied on never reading MockReadData could
not be safely read not being read.  The memory bots slow things down
enough that it was being read, and causing a crash.  Fixed the bad
data, and corrected the ordering of events.

BUG=130874
Review URL: https://chromiumcodereview.appspot.com/10704042