Revert 145993 - [Sync] Add location parameter to SyncChange in Sync API.
SyncChanges now require the location they were created as a parameter. This
is then used in the case that the change violates and invariant. We display
the location and type of erroneous change in about:sync, in addition to
disabling the offending datatype.

TBR=kalman@chromium.org, mnissler@chromium.org, isherman@chromium.org
BUG=135108
TEST=unit_tests


Review URL: https://chromiumcodereview.appspot.com/10766016

TBR=zea@chromium.org
Review URL: https://chromiumcodereview.appspot.com/10698141