Fix a renderer crash when dragging quickly over tab contents.
We were expecting to receive GTK+ events in the following order:
  drag-motion
  drag-data-received
  drag-leave
but in some cases, we don't get the drag-data-received event (e.g.,
if the user is moving the mouse quickly). If that happens, don't send
a DragLeave event to the renderer since it doesn't think we're in a
drag.

BUG=129446
TEST=Open 2 chrome windows, quickly drag a file from the desktop over both
  windows. This should not crash.


Review URL: https://chromiumcodereview.appspot.com/10704144