Remove app shortcuts when app is uninstalled on Linux.
To support this, shortcut creation on Linux for extensions has been modified so that the filename encodes the extension ID and the profile. Also, when creating shortcuts any existing shortcuts are removed first.

Web page shortcuts are not affected.

BUG=130456
TEST=Test uninstalling apps removes their shortcuts; test uninstalling
apps is not broken in any way; test shortcuts for web apps
are not broken in any way.


Review URL: https://chromiumcodereview.appspot.com/10698114