Merge 144898 - Implement CreateFile() for GData file system.
MERGE NOTE: This is slightly different from the original patch, and
was not merged with drover.  The difference is using GURL instead of
FileSystemURL in the argument, since FileSystemURL was introduced in a
later webkit change that we don't want to merge.  Functionally they
should be the same.

This CL is the last piece for implementing the File API
getFile(..., {create: true}, ...) for GDataFileSytem.

Currently the file creation is implemented as uploading of
an empty file to the server. The Drive server assigns the
resource_id, and further read or write operation just
treats it as a non-pinned file (e.g., when it is read or
written it is downloaded to GCache/v1/tmp/<resource_id>.<md5>).

BUG=132968
TEST=Manual test.
1) Be sure prerequisites patches (10636053, 10656043, and 10556003) are applied.
2) Edit image files on Google Drive folder by photo editor, and with "overwrite original" unchecked)
3) Verify a new file is created on the Drive.

Review URL: https://chromiumcodereview.appspot.com/10568012
Review URL: https://chromiumcodereview.appspot.com/10693150