gdata: Remove a werid behavior from GDataCacheMetadata::UpdateCache()
Though it's not documented in the function comment, UpdateCache()
had ability to remove an existing cache entry. This is more like a
surprising behavior. This patch removes the behavior.

Along the way, rename the function to be more descriptive.
Also rename RemoveFromCache() to RemoveCacheEntry().

BUG=136921
TEST=out/Release/unit_tests --gtest_filter=GData*

Review URL: https://chromiumcodereview.appspot.com/10690154