Merge 143644 - Try to fix a crash in the var tracking.
I think what is happening is that the list is mutated as we iterate over it.
This is actually quite a resonable thing to do: we're deleting objects, and
those objects may hold refs to other objects in our list.

BUG=http://crbug.com/133951

Review URL: https://chromiumcodereview.appspot.com/10633019

TBR=brettw@chromium.org
Review URL: https://chromiumcodereview.appspot.com/10702192