Fixing the way history service factory access the history (i.e., when it creats and when it doesn't.
Reviewers are TBR'd because they've already looked at this in the previous version of the CL (http://codereview.chromium.org/10536230/). The previous CL was reverted to diagnose Mac perf problems, but did not seem to be the cause. So putting back those changes in this CL and a couple others.

BUG=133762, 97804
TEST=unittests 

TBR=sky@chromium.org, dominich@chromium.org

Review URL: https://chromiumcodereview.appspot.com/10700193