Revert "Revert 146262 - Revert "Revert 142517 - [Sync] Refactor sync configuration logic.""
Relanding with support for purging partial types as part of SyncManager
initialization. In addition, the previous version mistakenly assumed it
could perform the DOWNLOADING_NIGORI state in SBH on normal restarts.
This is not the case. We bring back the condition to skip the nigori
downloading if HasSyncSetupCompleted is true.

original review at: https://chromiumcodereview.appspot.com/10483015

BUG=129665, 133061, 133219
TEST=unit/integration tests


Review URL: https://chromiumcodereview.appspot.com/10780002