Retire pulling down software_rendering_list.json from deps
We already checked in the file in src/.  In theory we should retire the pulling down in that same patch.  But at the time it seems to confuse grd and it fails a few bots.

This is another try.  Hope after a few weeks, all bots already picked up the new checked in version and we can safely remove the old one.

BUG=
TEST=tree

Review URL: https://chromiumcodereview.appspot.com/10790002