Add DPI information to video packets.
The intention is that capturers can add this information and the client can
make use of it to up-scale the desktop if it DPI is larger than that of the
host by an integer multiple. For example, a standard-DPI host viewed on a
retina-display Macbook Pro would no longer look tiny.

BUG=135089
TEST=EncoderVp8Test.TestDpiPropagation


Review URL: https://chromiumcodereview.appspot.com/10736046