Clean up WATCHLIST
Most uses of the Kleene star in this file were doing nothing, and only contributing to making it harder to read. I've removed all the Kleene star use cases that I found confusing, and a comment at the top of the file to remind people not to reconfusify the file.

R=nirnimesh@chromium.org
BUG=None

TEST=


Review URL: https://chromiumcodereview.appspot.com/10698159