Enable renderer side mixing behind a flag.
Lands the final bits required to turn on renderer side mixing behind
a flag!  Switching <audio> over to the PCM_LOW_LATENCY track to boot.
Specifically the flag is --enable-renderer-side-mixing.

The big picture: RenderThreadImpl owns an AudioRendererMixerManager
which RenderAudioSourceProvider uses to create AudioRenderMixerInput,
which extensions of AudioRendererSink.  When ARMM creates ARMI it
passes in callbacks which allow ARMI to retrieve an AudioRendererMixer
from ARMM once ARMI::Initialize() has been called.

When ARMM gets a request for an ARM from ARMI::Initialize() it will
first check if one exists and if so, hand out a reference.  If not, it
will create a new ARM instance (which it owns) and initialize it with
the proper hardware output AudioParameters and an AudioDevice based
AudioRenderSink.

ARM will immediately call Initialize() and Start() on AudioDevice which
will then begin requesting data via RenderCallback from ARM.

If the hardware sampling rate is not equal to the input sampling rate,
ARM will instantiate a MultiChannelResampler instance to resample the
audio as required for use by a PCM_LOW_LATENCY AudioDevice.

After ARMI::Initialize() nothing really happens until RASP() eventually
calls ARMI::Start() which will then register the ARMI instance with ARM.
At which point ARM will start mixing it into the output stream it's
providing to AudioDevice.

Everything works as far as I can tell, but an optimization pass and more
careful testing is necessary on Windows before the feature comes out from
behind the flag.

BUG=133637
TEST=New unit tests. Manual testing.

Review URL: https://chromiumcodereview.appspot.com/10636036