Don't turn on accessibility automatically on Linux.
With this change, a Linux user can still manually switch on
accessibility support with the --force-renderer-accessibility flag,
but it will never get switched on automatically. Let's merge this to
M21 so it never happens to users who don't want it, then come up with
a smarter way to detect automatically later.

BUG=137537
TEST=Run Chrome with "assistive technology" flag on, observe no slowdown in gmail anymore.
TBR=ben

Review URL: https://chromiumcodereview.appspot.com/10778040