gdata: Remove GDataRootDirectory
Originaly, GDataRootDirectory was introduced as a place to hold some
extra metadata like the largest changestamp etc. but we no longer
need GDataRootDirectory, as we now have a container called
GDataDirectoryService.

We'll remove GDataRootDirectoryProto once the migration to the new
metadata storage is complete.

BUG=135312
TEST=out/Release/unit_tests --gtest_filter=GData*; also confirmed that the existing proto was loaded properly

Review URL: https://chromiumcodereview.appspot.com/10808012