Fix bug where pyauto-based Gmail scroll test would not scroll.
I believe this test was broken by some revisions done in
https://chromiumcodereview.appspot.com/10575028.  The problem was that
in the event of scrolling in Gmail, the computed amount to scroll by was
always 0.  This change uses a scroll by amount of 1, allowing the scrolling
to work again in Gmail.

BUG=136554
TEST=Verified the Gmail scroll test works on my local linux machine with this change.
NOTRY=true


Review URL: https://chromiumcodereview.appspot.com/10807030