gdata: Stop saving GDataFile::kind_ in proto
kind_ is only used for computing stats when documents list feed is
loaded from the server, hence there is no need to save.

Besides, saving C++ enum in proto is fragile, as C++ enum can change
very easily which causes incompatibility with the enum values saved
in proto. If we want to save enum in proto, we should define it in
the proto file instead.

BUG=135162
TEST=compiles

Review URL: https://chromiumcodereview.appspot.com/10806033