Fix 2 ChromeDriver compatibility crashes.
ChromeDriver compatibility was broken with 145212 because the BackwardsCompatAutomationMessageFilter expected all outgoing messages to be SendJSONRequest, whereas 145212 changed it to SendJSONRequestWithBrowserHandle. The simplest fix would be to just change the expectation, but instead I've gone ahead and migrated ChromeDriver to use SendJSONRequest, so that SendJSONRequestWithBrowserHandle can be deprecated easily.

ChromeDriver was crashing on close on windows with older Chrome's because ProxyLauncher's TerminateBrowser member function was calling quit on the first browser index. We don't allow any outgoing messages except SendJSONRequest. I simplified shutdown to just kill the ipc channel, which should cause chrome to terminate.

BUG=chromedriver:99
TEST=none


Review URL: https://chromiumcodereview.appspot.com/10780022