Initial commit of the App Host stub.
This stub will be installed at user level when a user installs Chrome apps. Shortcuts to those Chrome apps should point to the stub, rather than to Chrome. The stub dynamically locates a chrome.exe installed at either user- or system- level, and will eventually even trigger an installation of the Chrome binaries (not Chrome itself) if they are not installed. Once chrome.exe is located, it is launched with the same command-line flags as originally passed to the app host.

BUG=None
TEST=execute app_host.exe, confirm it launches Chrome.

Review URL: https://chromiumcodereview.appspot.com/10559090