sync: Remove SyncInternal
Also removes unnecessary functions (like UpdateCryptographerAndNigori).  An additional bonus of removing this is that we can remove the ForTest* functions (previously they depended on SyncInternal behavior which wasn't defined in headers). I already had a patch in the works to remove those, so I'll upload that part separately.

This is based off of http://codereview.chromium.org/10701046/

BUG=137740
TEST=sync_unit_tests, unit_test

Review URL: https://chromiumcodereview.appspot.com/10810020