Disable blurring for extension bubbles
Extension bubbles in HiDPI mode were drawing incorrectly.

The problem was that the blur filter we placed on bubble windows was interfering with OpenGL drawing.

I've filed an Apple bug about this:
  http://openradar.appspot.com/11920246
In the meantime this CL disables the blur filter for extension bubbles.

BUG=135382
TEST=


Review URL: https://chromiumcodereview.appspot.com/10808046