Implement installation of the Chrome App Host.
The Chrome App Host is a simple exe that delegates to a Chrome Binaries installation at user or system level. If no installation is available, it will trigger an installation.

The Chrome App Host prevents the Chrome Binaries from being uninstalled, except in the case of an upgrade to system-level.

The Chrome App Host is implemented in this uncommitted CL: http://codereview.chromium.org/10559090/ .

BUG=None
TEST=None

Review URL: https://chromiumcodereview.appspot.com/10665002