isolate: Use a temporary directory on the same file system than root_dir.
Reuse the same logic than the one used in run_test_from_archive.py. Otherwise
this makes the run and remap command to be really slow when $TEMP is on a
separate file system since they need to copy all the files. This happens when
tmpfs is used or when another partition than C:\ is used on Windows to store
the sources.

TBR=cmp@chromium.org
NOTRY=true
BUG=
TEST=


Review URL: https://chromiumcodereview.appspot.com/10790107