gdata: Make GDataFileSystem::ReadDirectoryByPath() much more efficient
Previously, ReadDirectoryByPath() was returning DataDirectoryProto,
which contains all child files and directories, recursively.
This means, calling ReadDirectoryByPath() on the root directory is
super expensive as the returned GDataDirectoryProto contains the
entire tree as proto. This was super expensive.

With this change, ReadDirectoryByPath() will only return contents
of the target directory as a flat vector.

BUG=137871
TEST=out/Release/unit_tests --gtest_filter=GData*; also confirmed that the file manager worked as before, including the "save as file" to Drive

Review URL: https://chromiumcodereview.appspot.com/10815008