Make use of -lppapi explicit to allow for correct library order.
This CL removes the implicit -lppapi and pthread in the Makefiles.  Since ppapi lib 
does not exist on the host (on NaCl it's actually a linker script), we need to 
create a bogus ppapi.lib for host Pepper development.  In addition, on Windows we
do not have a pthread library so we create a mini pthread library to support the
examples.


This CL also fixes tries which are broken on Windows.

Review URL: https://chromiumcodereview.appspot.com/10815039