Tidy up Android's is-tablet functionality.
Remove the code in sync/ that checks if the device is a tablet
and perform the check in the chrome/ layer based upon a
command line switch.

Additionally, rather than looking for a tablet when (not) setting
the Mobile specific user agent, look for --use-mobile-user-agent. The
embedder should set these command line switches as appropriate, so
remove the code that used to set them in the generic process startup.


Review URL: https://chromiumcodereview.appspot.com/10808052