Merge r147634: Fix checking for implied permissions in PermissionSet
Checking for implied permissions (e.g., kWebRequest => kWebRequestInternal)
was previously done only in the constructor parametrised by an extension.
Since the permissions should be implied independently of such context,
this applies the check in situations when the extension is not known.

Browser-test is being considered, but not written yet. If that happens,
it will be on a separate CL.
BUG=136103
TEST=Install the extension referenced by the bug, go to Options, click Allow, and observe no error in the background-page console log.
Review URL: https://chromiumcodereview.appspot.com/10805088