shell dialogs: Move the linux implementations.
This moves the gtk and kde implementations to a common directory so a search will get all the code when we need to remove the normal gtk port.

This isn't a pure move. Parts of SelectFileDialogImplKDE had to be rewritten to not use BrowserThread.

BUG=134529
TEST=none

Review URL: https://chromiumcodereview.appspot.com/10790083