JSON Schema Compiler supports functions as PropertyTypes.
Before, the compiler raised an exception when a function was used as a property,
which happens whenever a function is passed as a parameter.

The solution presented here is to create a bool has_<function_name> to allow
the hand-written C++ code to know whether or not the function was passed in.

BUG=138850
TEST=function_as_parameter


Review URL: https://chromiumcodereview.appspot.com/10824002