HiDPI: Scale viewport for pepper flash fullscreen
Pepper flash fullscreen mode was not rendering correctly in HiDPI mode. The problem was that we weren't taking into account the display scale factor when setting the view port.

This CL pipes the screen info to RenderWidgetFullscreenPepper so that the display scale factor can be used.

BUG=130323
TEST=

Review URL: https://chromiumcodereview.appspot.com/10807039