Add .revisions() implementation to *Checkout classes.
This permits to count the number of check-ins between two revisions. This will
be used by the CQ to determine if a try job 'expired', e.g. it is so old that it
can't be used anymore. The .revisions() function specifically count the number
of commits, instead of just doing arithmetic on the svn revision numbers.

R=cmp@chromium.org
BUG=
TEST=Manually:
  import checkout
  s = checkout.SvnCheckout('/path/to/chrome/src', None, None, None, 'svn://svn.chromium.org/chrome/trunk/src')
  s.revisions(148323, 148330)
  s.revisions(148323, None)
  g = checkout.GitCheckout('.', None, 'master')
  g.revisions('HEAD^^^^^^^^', None)
  g.revisions('HEAD^^', 'HEAD')


Review URL: https://chromiumcodereview.appspot.com/10821011