Add GL surface creation for Android.
This creates either a texture transport surface (for compositor use) or a native window surface.
    
Also add a way to look up a surface handle that cannot be serialized over IPC. We simply get it from the tracker as we are running the GPU thread in the browser process.

BUG=136923

Review URL: https://chromiumcodereview.appspot.com/10795058