Remove details from BROWSER_CLOSING and BROWSER_CLOSED notifications.
The details was a boolean indicating whether this is the last browser window closing. As part of this change the logic for saving pinned tabs has been changed to not use the boolean in the details, and has been updated to handle the case where a user exit or closing of the last window is cancelled by an onbeforeunload event.

Also as part of this change, both the pinned tab and multi profile features have been updated to handle cases of the browser process not being closed due to background mode or (in the future) packaged apps, and then more browser windows being opened.

BUG=None
TEST=Test shutdown 


Committed: http://src.chromium.org/viewvc/chrome?view=rev&revision=148281

Review URL: https://chromiumcodereview.appspot.com/10800031