browser_tests.isolate: fetch whole ../third_party/WebKit/Tools/Scripts/webkitpy/
Do not individual files that are needed since the list of needed files for
new-run-webkit-websocketserver is too often changing.

Contrary to the previous .isolate CLs, this one was done manually. A rule in isolate_common.py will be added in a follow up change.

TBR=dpranke@chromium.org
NOTRY=true
BUG=139426


Review URL: https://chromiumcodereview.appspot.com/10828059