PrefValueStore and PrefService expose the value of a pref returned by the highest-priority PrefStore only. Recommended values are interesting even if they are currently being overridden by a higher-priority source.
This CL adds a getter for the recommended value and makes pref change
notifications be sent on all value changes, even if the change occurs in
a PrefStore that is currently overridden. This will allow the settings UI
to adapt when recommended values change.

BUG=104955

Review URL: https://chromiumcodereview.appspot.com/10821047