Fix segfault in JNIEnv function hooks for test
On newer JB builds, the JNINativeInterface struct appears to be held
in read-only memory so attempting to repoint individual function pointers
in it causes segfault. So instead we repoint the top-level struct pointer
to refer to our own copy held in read/write memory.

BUG=137348
TEST=JNIAndroidTest


Review URL: https://chromiumcodereview.appspot.com/10828050