Extensions Docs Server: Proper $ref handling
If an API has a $ref type to another file, it is now displayed with the prefix
of the API it is a ref to. For example if Tabs references Windows, the ref will
be windows.Window, but if Tabs references Tab, the ref will be just Tab.

Also, $refs are handled in descriptions.

BUG=131095

Review URL: https://chromiumcodereview.appspot.com/10826024