Modifications to get skia headers compiling on iOS.
This is a barebones implementation of the skia target in skia.gyp.  iOS will not
use a full skia implementation, but it will use an occasional header here and
there.  Those headers require an implementation of SkDebugf_FileLine(), so this
CL adds a barebones skia target that simply compiles skia/ext/google_logging.cc.

BUG=None
TEST=None


Review URL: https://chromiumcodereview.appspot.com/10832081