Fix SyncManager initialization failure crash.
This crash was introduced in r148926, which changes SyncManager init
behaviour.  Previously, the SyncManager would end up with a valid
SyncScheduler regardless of whether or not the initialization failed.
This SyncScheduler would later play an important role in shutting down
the syncer.  Without a scheduler, the shutdown process crashes.

The ProfileSyncServiceTest.CorruptDatabase test simulates the scenario
we fail to load a sync directory.  This will result in a SyncManager
initialization failure.  It is intended to test that this failure is
handled by the ProfileSyncService, SyncBackendHost and SyncManager.

BUG=139723


Review URL: https://chromiumcodereview.appspot.com/10830100