ContentShell rendering support on Android.
This renders to a SurfaceView using texture sharing on the GPU
thread. There is a bit of a public interface added to let the
app create a cmdbuffer (we want run GL contexts on the same thread
when sharing resources on Android) and perform the handshaking
needed for sharing the two compositor buffers.

It has bugs, such as that it doesn't handle it yet when the UI
surface handle changes (e.g. when hiding and restoring the app).

BUG=136923


Review URL: https://chromiumcodereview.appspot.com/10823051