sync: Improve tracking of useless updates
Our counts of non-useless updates are wrong in part because of the way
we track deletions.  When an item is deleted, we flip the IS_DEL bit.
On restart, we purge all fully synced IS_DEL items from our directory.
However, we might still receive some re-deliveries of the tombstone for
that item.  The old algorithm would consider tombstones of non-existent
items to be "useful", despite the fact that they're often just
reflections of updates we've already received.

The updated algorithm isn't perfect.  There's no easy way to distinguish
between a tombstone created by another client where this client never
saw the original (which one could argue is useful, and definitely not a
reflection) and the re-delivery of a tombstone created by another client
which this client has already received and processed long ago (which is
definitely not useful).  Both scenarios look the same, and both will be
counted as 'reflections' under the new algorithm.

A similar issue arises with UNIQUE_CLIENT_TAG items.  When we delete
them we force their version to 0 locally.  This makes our regular
reflection detection algorithm ineffective.  This change updates the
algorithm to consider tombstones to items that have UNIQUE_CLIENT_TAG
and are locally deleted to be reflections.  It's not perfect, but it
should be good enough for our purposes.

BUG=139684

Review URL: https://chromiumcodereview.appspot.com/10834095