Add ability to format errors as a list of temp-allow rules to paste into a DEPS file.
The idea behind this change is that the tool now lets you do this:
a) Update a DEPS file for some directory with the intended rules for
the directory, e.g. rules disallowing certain directories (say
"chrome/browser") and allowing others (e.g. "chrome/browser/apis").
b) Run this tool on that particular directory to generate a list of
temp-allow rules.
c) Paste the temp-allow rules into the DEPS file for the directory.
d) Work on eliminating dependencies until the temp-allow rules list in
the DEPS file reaches zero.

To accomplish this change, and fix the bug that existed that the
checkdeps tool didn't work correctly if you specified a start
directory, I refactored the code to always use the approach I had
introduced in CheckAddedCppIncludes, which when you request rules for
a given directory generates all parent directories' rules first, up to
the base directory.

Also, to enable better extraction of error information for the
eventual formatter, I externalized formatting to separate classes
and had the code collect only the logical error information.

While in there, fixed use of quotes to match coding standard.

BUG=138280

Review URL: https://chromiumcodereview.appspot.com/10832062