Add a modified LLVM/Clang compiler to skim allocated types and sizes.
This package consists of a locally modified version of Clang.  It instruments
code around 'new' and 'delete' expressions to skim allocated types and sizes.

The difference from the original Clang is in allocated-type.patch.

It is only for debugging and memory profiling, not for building daily-use Chrome.


BUG=123758
Review URL: https://chromiumcodereview.appspot.com/10834103