gdata: Introduce GDataWapiFeedLoader class
This class is introduced to separate code related to WAPI
(codename of Documents List API) feed loading.

To minimize diffs, the the class is defined in gdata_file_system.cc
The code will be moved to a set of new files in a separate patch.

BUG=130669
TEST=out/Release/unit_tests --gtest_filter=GData* && out/Release/browser_tests --gtest_filter=GData*

Review URL: https://chromiumcodereview.appspot.com/10828126