Android: move various gyp defines out of envsetup.
These gyp defines are set unconditionally in envsetup, which means they
are not in fact part of the "environment" but just the defaults for the
Android build. Move them into common.gypi instead to make this clear.
Several of them were already the correct setting and can simply be
removed.

The makefiles generated by gyp are identical before and after this
change.


Review URL: https://chromiumcodereview.appspot.com/10825131