Changes ProviderImpl to use a single DataFetcherFactory instead of a vector of DataFetcherFactories
In the current implementation, any OS can only use at most one DataFetcher, so
there will always be 0 or 1 DataFetcherFactories in the vector. Therefore it
is simpler to use a single DataFetcherFactory instead of an array. Since DataFetcherFactory is a pointer itself, it can be NULL when there are no available DataFetchers.

BUG=none


Review URL: https://chromiumcodereview.appspot.com/10837055