Fix sourceAbort() for BMFF streams.
MP4StreamParser::Flush() (called eventually by sourceAbort()) did not reset the
parser state, and so would only work if the call happened to come when the
parser was already in a particular state (which happened to always be true for
the browser test harness it was originally tested with).

Additionally, the format-specific buffer conversion methods did not cause
an error to be propagated to the caller on failure, resulting in unexpected
success for the added test.

BUG=
TEST=MP4StreamParserTest.TestFlush


Review URL: https://chromiumcodereview.appspot.com/10843044