Add parsers for Drive v2 Change/ChangeList/Parents
Parent means a folder containing a file in Drive v2 API.
If a file does not contain any reference to its parent, it should be
interpreted as the file is under the root directory.
Changed existing unit test for checking the case that parent is missing.

BUG=chromium:127728
TEST=unit_tests --gtest_filter="DriveAPIParserTest.*"

Review URL: https://chromiumcodereview.appspot.com/10829139