Make the getAttention badge grey.
The grey badge fades in using the existing animation, and when the
user clicks on it, it snaps to the full-color version, without an
animation.

Also avoided granting access on right clicks, which I noticed because
of the color change.

This change has the side-effect that pageaction.show() will cause the
badge to fade in.

This doesn't yet show the grey background, as that takes
platform-specific code.

BUG=133142


Review URL: https://chromiumcodereview.appspot.com/10823143