[Chromoting] Use different talkgadget prefixes for client and oauth.
With this change, Chromoting will use chromoting-oauth.talkgadget.google.com
to talk to the talkgadget during the auth phase and will use
chromoting-client.talkgadget.google.com when the client is interacting with
the talkgadget.

Host interactions with the talkgadget will be controlled by policy and will be
handled in a separate cl.

BUG=132345
TEST=None


Review URL: https://chromiumcodereview.appspot.com/10837106