Fixing crash in V8ValueConverter.
It appears that the crash occurs when the V8ValueConverter is passed an object
that has a named callback (v8::Object::HasRealNamedCallbackProperty()) that
is an internal property (something that is not defined in JS, but intercepted by
the c++ side).

Something like a DOM input element is a good example of this.
The crash doesn't happen with every object that has internal properties and
named callbacks, but so far I am unable to figure out how to programatically
differentiate between the ones that will crash and the ones that won't.

My current solution is to not convert any named callback.
BUG=139933


Review URL: https://chromiumcodereview.appspot.com/10837066