Improve accessibility of user image selection screen.
Gives each image a spoken description. Fixes the ARIA markup so it's
exposed as an interactive listbox when accessibility is turned on.
Also, remaps up/down to be the same as left/right when accessibility
is on, since the 2-dimensional structure is not exposed.

Here's exactly the feedback it gives now, when you tab to the list box
and one item is selected:

"Choose a picture to display for your account on the sign-in screen.
 Cupcake, List box 7 of 20".

BUG=126186,chromium-os:26999
TEST=manual testing with spoken feedback on

Review URL: https://chromiumcodereview.appspot.com/10376003