Merge 150155 - Add the screen offset to the location of the flash context menu.
This fixes a bug where the context menu would display on the wrong monitor if the plugin was in full screen and not on the primary monitor. The coordinates to |ShowContextMenu| are expected to be relative to the window position. However, the coordinates given by |windowRect| includes the screen offset so we need to correct for that.

BUG=139641
TEST=Tested context menu on win flapper on both the primary and secondary monitors. Mac context menus use the mouse coordinates. Linux fullscreen for multiple monitors is currently broken.
TBR=brettw
NOTRY=true


Review URL: https://chromiumcodereview.appspot.com/10837117

TBR=raymes@chromium.org
Review URL: https://chromiumcodereview.appspot.com/10834197