IndexedDB browser test for migrating from leveldb schema v0 to v1.
It doesn't check that the integer version was actually added
to leveldb though, which is unfortunate.  It ensures that
opening a v0 database doesn't fail an assert, which is what
would happen if you opened an unmigrated database with code
that expects to find a stored integer version.

BUG=108223

Review URL: https://chromiumcodereview.appspot.com/10826159