Switching nacl.sdk.mono entry point back to something in-repo.
Previously the annotator script would select between entry points based
on the name of the builder passed to it by the annotator.
In refactoring it, this behavior was lost.

The old behavior is only relevant to the nacl.sdk.mono waterfall.
Pointing that waterfall back at the old script
(as it is a non-chromium style waterfall).

BUG=None
TEST=None
R=cmp@chromium.org

Review URL: https://chromiumcodereview.appspot.com/10825228