Roll Android SDK in android_tools repo to api version 16.
The install-build-deps-android-sdk.sh is no longer necessary, since the SDK is checked out as part of the tree.  It is still useful as an easy way to install the SDK/NDK if we want to use an alternate (i.e. if we explicitly export ANDROID_SDK_ROOT it won't use the tree version).

BUG=


Review URL: https://chromiumcodereview.appspot.com/10823174