crbug.com/124865 - DCHECK failure when pinning Chrome Web Store to Launcher
CrOS Aura: Fix pinning Chrome Web Store to Launcher

When requesting to load the extension icon, we now check whether the ext. resource is empty. If so, we request DONT_CACHE (instead of CACHE). Caching fails two DCHECKS when the ext. resource is empty. Also the key involved in caching would be an empty string when the ext. resource is empty.

Earlier solution, later abandoned (reverted):
Icons are not required in extension or app manifests. Modified
Extension::HasCachedImage and Extension::SetCachedImage to not fail
a DCHECK nor to cache an image, if the ExtensionResource is empty.

See also the now obsolete: http://codereview.chromium.org/10823115/

BUG=124865
TEST=Click on Apps icon in the Launcher;
     Right click on Chrome Web Store;
     Select Pin to Launcher;
     Observe no DCHECK on Debug build on linux desktop

Review URL: https://chromiumcodereview.appspot.com/10830123