Make java.gypi export jar path to input_jars_paths variable
Both java.gypi and apk_test.gypi expect the jar path in
input_jars_paths. This means that we were specifying the dependence in
both 'dependencies' and 'input_jars_paths'. This change makes it so that
we don't need that redundancy.

Also, make java.gypi use input_jars_paths in its input so that we
actually rebuild targets when we should.

BUG=136756


Review URL: https://chromiumcodereview.appspot.com/10837143