Don't reset the layout on RenderText display rect width change.
Previously, I thought width was used in layout in render_text_linux.cc,
but it turns out that the width that's passed in to SetupPangoLayout()
gets cleared right after. So the width of the display rect does not
actually affect layout and thus does not need to cause the layout to
be reset.

BUG=137064
TEST=No behavior change.

Review URL: https://chromiumcodereview.appspot.com/10829261