Add the ability to set the upstream branch
I can never remember the syntax (and always forget I have to specify the branch to set the upstream branch on) when using "git branch".

This patch just allows git-cl upstream to set the branch if an argument is provided. With no arg it prints the current one as it does today.

Review URL: https://chromiumcodereview.appspot.com/10825285