Fix extension link error on Android.
The USB extension binding code introduced in r134423 is excluded from
the build when OS==android, but still calls the excluded code. None of
the other bindings are excluded on Android currently, and the files do
actually build, so just reinclude them for now. More extension code may
be excluded from android in future, but in a less piecemeal way.

Review URL: https://chromiumcodereview.appspot.com/10855047