Partial revert of 146621
The original patch introduced a crash (GetCompositor() can return NULL), but
reverting the whole thing is not possible without a lot of other reverts, and
the proper fix is fairly involve (likely not reasonable to patch to the branch).

This only reverts the problematic part, and a follow up will do a proper fix.

BUG=141756


Review URL: https://chromiumcodereview.appspot.com/10828262