Fixing bug in updating file list in save-as dialog.
Problem occuired because file manager code uses the selectedIndexes property of selectionModel. But setter for this methos is implemented only for multi-selection list, but not fo single selection (which is used in the save-as dialog). It fails to restore selection and consequently fails to restore lead item. Resetting lead item stops renaming.

BUG=138975
TEST=Manual test


Review URL: https://chromiumcodereview.appspot.com/10823201