Remove unnecessary assignment in transport_client_socket_unittest.cc.
This is part of Chrome for Android upstreaming.

This line was introduced by r137387 which turned |sock| into a scoped_refptr
rather than a raw pointer (that needed to be initialized). This assignment is
obviously not needed anymore.

Internally we don't have this extra '= NULL' which causes an unnecessary diff.

BUG=136750


Review URL: https://chromiumcodereview.appspot.com/10855123