Merge 150979 - Fix Ash notification updates
In Chrome, it turns out that every Notification has a unique id. In the Ash implementation we need to track and update that id when an existing Bubble is updated.

This also ensures that the UI is safely created initially (in case the initial asynchronous Update is delayed) to prevent crashes and to make Layout more consistent.

Also includes some cleanup of the WebNotificationTray API.

BUG=141285


Review URL: https://chromiumcodereview.appspot.com/10855079

TBR=stevenjb@chromium.org
Review URL: https://chromiumcodereview.appspot.com/10830283