Make jars build from a single ant .xml template
The ant build files for each of the jars that we create consisted of
only fairly straightforward boilerplate. With the recent changes to
java.gypi and java_aidl.gypi we can now build these correctly from a
single .xml with GYP passing in defines for the minor differences.

BUG=136756


Review URL: https://chromiumcodereview.appspot.com/10831227