Fix lifecycle-vs-threads of GpuVDAContext3D.
This context3d is meant to be created on the render thread but can only be
dereferenced on the compositor thread.  Before this CL this was violated as
RenderThreadImpl::GetGpuVDAContext3D() checked for context loss to decide
whether to recreate the context.  Now context loss proactively notifies the
render thread, instead.

BUG=142047


Review URL: https://chromiumcodereview.appspot.com/10832264