Fix white flash when user signs out and fix wallpaper animation regression
When user signs out and transit to the login screen, the wallpaper fade in from a white screen. Described in 141563, this Cl fixed it.
Also, after user login, or switch users at login screen. when wallpaper change, the new wallpaper should fade in nicely. There is a regression in recent builds. This fix fixed the regression as well. 

BUG=141563, 141676, 142042

Review URL: https://chromiumcodereview.appspot.com/10827284