Set TCP_KEEPALIVE idle and interval on Android.
This is part of Chrome for Android upstreaming.

It affects how long our side of an idle TCP socket can know the other end has
closed the connection. The default value is 2 hours which is too long. We
should use the same values on Linux.

BUG=136750


Review URL: https://chromiumcodereview.appspot.com/10831286