Support multiple disable reasons in ExtensionPerfs
Use a bitmask to represent all the reasons an extension is disabled.
Switch to a new key in ExtensionPrefs for the bitmask.
When we init ExtensionPrefs, check for the old key, migrate the reason over
to the new key, and delete the old key.

BUG=142762


Review URL: https://chromiumcodereview.appspot.com/10828316