Add IPC::Sender and GetRoutingID() to WebContents for convenience and safety.
Code that is or has a WebContentsObserver can already use safe
versions of Send() and GetRoutingID() that perform NULL checks and
delete the Message object and return MSG_ROUTING_NONE when
appropriate.  This brings the same thing to code that has just a
WebContents.

BUG=none


Review URL: https://chromiumcodereview.appspot.com/10830337