Search state transitions not working
This change codifies and fixes the state transitions used in
the Search UI.  It documents the transitions in tabular
form in search_types.h detailing the user interaction that
results in either animated or non-animated transitions.
The pending navigation signal is removed, so all transitions now
occur on the navigation "committed" signal.

BUG=133529
TEST=Manual
R=sky@chromium.org, kuan@chromium.org


Review URL: https://chromiumcodereview.appspot.com/10837240