Propagate Aura gesture timestamps to WebGestureEvents
We do this for all other input events, it was probably just missed for gesture
events.  The gesture recognizer appears to use a usefull timestamp value in
most cases (the time of the last touch which is what we want except perhaps for
timer-based gestures like long press).

BUG=143172


Review URL: https://chromiumcodereview.appspot.com/10831356