Refactor browser window zoom handling and enable zoom icon on all platforms.
On all platforms:
After this CL there should be zoom icon in the omnibox when on a page with non-default zoom (- for less than default, + for more than default). If the zoom is reset with Ctrl+0, no bubble should show up and the icon should disappear. If tabs with different zooms are switched between, the zoom icon should update itself for each tab without showing a bubble.

On Views and GTK:
While changing zoom with Ctrl +/-, a zoom bubble should show up if you zoomed *to* a non-default level. If you quickly change to a different tab after changing the zoom level, the bubble should disappear.
 
BUG=86817,128817
TEST=Compiles, passes tests, doesn't regress current zoom functionality.
TODO=Zoom *bubble* on Mac.
NOTRY=true

Review URL: https://chromiumcodereview.appspot.com/10736028