Fix user agent override restoration pipeline
There were some disconnects in the pipeline that prevented the user agent override
string from being saved and restored from the temporary tab structions in sessions/,
but couldn't be tested on Android because the pathways weren't being exercised in
the same way.  This CL finishes hooking it up for ChromeOS (and presumably the other
platforms).
    
BUG=140600


Review URL: https://chromiumcodereview.appspot.com/10827168