Advertise a minimum alignment for SharedMemory::Map().
No real changes, just adds a DCHECK() and unit test for a minimum
alignment that callers can count on from Map().  Map() is already
returning addresses which are aligned to the platform page size.

Necessary for a media use case in which a wrapped shared memory
object is passed around and is expected to have a certain alignment.

BUG=none
TEST=unittests, try bots.


Review URL: https://chromiumcodereview.appspot.com/10827400