Add GPU memory tab to the task manager.
Each renderer process will have all of the resources that the GPU process has allocated on its behalf (including WebGL resources, compositor resources, and the backbuffer) included in its printed total.

The GPU process will have the all resources currently allocated by the GPU process in its column.  This will be approximately the sum of all other rows.  The row for the GPU process is printed in ()s to draw attention to the fact that its size includes duplicates from other processes' sizes.

I happy this UI scheme (having played with a few), but I'd wider feedback.

Note that we do not account for swapchains or for backbuffers allocated by the browser process, so those allocations are not counted in the total.

BUG=140157

Review URL: https://chromiumcodereview.appspot.com/10854076