Don't show the button for hidden page/browser actions or script badges.
Browser Action, Page Actions and Script Badges can be hidden, which causes us to get invalid |button| and |frame| (respectively). We'll need to figure out what to do in these cases (on all platforms) but the minimum for now is to ignore the keystroke to avoid a crash. 

BUG=140396
TEST=None
Review URL: https://chromiumcodereview.appspot.com/10828364