Move the gyp variables used for grit defines on android to actually work.
The enable_themes and enable_extensions gyp variables were in a different
scope that prevented the grit_defines from having the correct value when
evaluating the condition.  Moves the variable assignment into the same
scope as the variable declaration and now everything is happy.

TEST=touch chrome/browser/browser_resources.grd && \
     make chrome_resources (check about_version does not have theme icon)

BUG=


Review URL: https://chromiumcodereview.appspot.com/10828396