Remux bear-640x360.webm so it is a valid WebM file & update the ChunkDemuxer tests that rely on it.
FFmpeg generated the existing file with negative timestamps at the beginning which violates the WebM spec. 26ms had to be added to all the timestamps in the file to make it valid.  This slightly shifted the file layout and cluster boundaries  so ChunkDemuxer tests needed to be updated.

BUG=122913

Review URL: https://chromiumcodereview.appspot.com/10861030