Update run_slavelastic to Handle Network Disconnects.
Fixing an issue where the slave was unable to open the new zip file (I assume the network path was down for a second, but not sure) by catching the error and nicely aborting. This gives the tests that still need to be processed after a chance.

BUG=


Review URL: https://chromiumcodereview.appspot.com/10832341