[sync] Fix timestamp mechanism in CredentialCacheService
CredentialCacheService shares sync credentials between Metro and Desktop
Chrome on Win 8. While updating a cache, we write a timestamp to the
cache file, to make sure that credentials are only applied from a newer
cache (with the user's more recent sync settings) to an older cache, and
not vice-versa.

CCS was erroneously persisting a base::TimeTicks value as a timestamp.
It was also incorrectly handling cases where the alternate cache was last
written by an older version of chrome, and therefore, missing the "last
updated time" field.

This patch does the following:

1) Updates the timestamp mechanism of CCS to use base::Time. We now
   write a base::Time value to the "last_cache_update_time" field.
2) Deprecates the old "last_updated_time" field, which contains
   TimeTicks values.
3) Auto-heals caches written by older versions of Chrome, that contain
   the old timestamp field, or do not contain a timestamp field at all.
4) Updates the timestamp comparison mechanism to handle cases where the
   alternate cache does not contain the "last_cache_update_time" field.

BUG=143529,143214
TEST=Sign in, exit, delete the "last_cache_update_time" field and add a "last_update_time" field to  "Sync Credentials" and restart Chrome. The "last_cache_update_time" field must get restored, and any old instances of "last_update_time" must be removed.

Review URL: https://chromiumcodereview.appspot.com/10837328