Use mac_build.gypi to build breakpad utilities on iOS
This uses the generic support for bouncing iOS dependencies through
ninja to build breakpad utilities, as is done for other host-targetted
utilities needed for iOS builds.

The built utilities are copied to the non-ninja output locations, so that
the bots can find them the same way they do for a Mac build.

BUG=None

Review URL: https://chromiumcodereview.appspot.com/10855263