Add new TextInputState parameters to carry information from WebViewImpl::textInputInfo.
This changes ViewHostMsg_TextInputState to take a params struct, containing the existing
parameters plus the additional information provided by WebViewImpl::textInputInfo. That is, it
adds text value, selection offsets, and composition offsets to the parameters.

There are no functional changes as a result of this CL. The new parameters will be used to
implement Android IME functionality in future CLs.

BUG=136766
TEST=RenderviewImplTest.OnImestateChanged


Review URL: https://chromiumcodereview.appspot.com/10800014