win: Always present AcceleratedSurface via the UI thread.
Presenting from a present thread causes flickering issues when Aero is turned on.

Synchronization between the CPU and GPU still happens on the present thread because moving everything tot the UI thread caused jank on resizing.

The UI thread polls until DWM returns that there are no pending presents so that when it calls Present it won't block.

This did not work as well with Aero turned off so I used the original method of presenting on the present thread in that case.

BUG=120904

Review URL: https://chromiumcodereview.appspot.com/10823133