Update next_id when loading sync DB
The directory code is smart enough to maintain a bit of extra slack in
the next_id so we don't accidentally reuse IDs if we crash.  However,
it only works after we've saved changes to the directory at least once.

This change ensures that we save changes immediately after we load the
directory, so that we can recover from crashes that occur before we
would otherwise save changes due to natural causes.

This won't do anything to help users who've gotten into a bad state
because next_id was corrupted by an early crash, but it will prevent us
from corrupting databases in the future.

BUG=142987
TEST=See below:
Examine 'next_id' in sync DB.  Open chrome and close it within two to
ten seconds.  Verify 'next_id' has been incremented.


Review URL: https://chromiumcodereview.appspot.com/10832329