Removing unused RemoveObserver (ReleaseBlocker)
Background: This was a remainder from my initial attempt to re-use the tray bubble. Looking at the code I don't see a reason why I should remove the observer when I have never set it in the first place. I assume that usually no harm was done when the owning anchor buttons still existed upon destruction. However - when the button is gone it is obvious that that can't be good.

BUG=143716
TEST=visual


Review URL: https://chromiumcodereview.appspot.com/10867014