Relanding this with build fixes.
Revert 152995 - THIS BROKE THE BUILD.

Revert 152892 - Merge 152799 - Ensure that the window switcher in Chrome metro mode on Windows 8 works with themes. Currently it is non functional
as the tab strip bounds extend beyond the location of the switcher and the BrowserView in the WM_NCHITTEST
handling always returns HTCAPTION when we hover or click on the switcher button.

Fix is to ensure that the tab strip width with themes on is limited by the width of the switcher button. We also
handle the hit test query in the BrowserView to see if it is within the bounds of the switcher.

BUG=133070
R=sky
Review URL: https://chromiumcodereview.appspot.com/10872003

TBR=ananta@chromium.org
Review URL: https://chromiumcodereview.appspot.com/10878018

TBR=ananta@chromium.org
Review URL: https://chromiumcodereview.appspot.com/10870042

TBR=kerz@chromium.org
Review URL: https://chromiumcodereview.appspot.com/10870045