NaCl: Move NaCl's Breakpad tests into the Chromium repo
This imports tests/inbrowser_crash_test from the NaCl repo into the
Chromium repo with a couple of minor changes:

 * crash_dump_tester.py is changed to locate
   native_client/tools/browser_tester via a new relative path.

 * I've renamed it from "inbrowser_crash_test" to
   "breakpad_crash_test" to make it more easily findable when
   searching for code related to Breakpad, and because "inbrowser" is
   redundant for code in the Chromium tree.

   (I think I originally omitted "Breakpad" from the directory name
   when the test lived on the NaCl side because the test was agnostic
   about the crash reporter implementation -- no Breakpad-specific
   code lived in the NaCl tree.)

 * Update copyright messages to say "Chromium Authors".

Since the test tests Chromium, the Chromium repo is the right place
for it.  Having the test here will make it easier to debug the test's
flakiness on 32-bit Windows.

BUG=143413
TEST=nacl_integration

Review URL: https://chromiumcodereview.appspot.com/10873019