Obfuscated Gaia ID fetcher
As part of our Push Messaging API, we need to fetch the Obfuscated GAIA ID from the server.  This code fetches the Obfuscated GAIA ID, and uses it to construct a channel-id which is used in the push messaging process.  The Push Messaging client hands the channel ID to their push messaging server, which uses it when they communicate with the Google back end for push messaging.

BUG=139658

Review URL: https://chromiumcodereview.appspot.com/10836182