De-craxify locking in VaapiVideoDecodeAccelerator's Decode methods.
Previously, lock_ would be acquired and released in an ad-hoc manner, allowing
state_ to change between locked sections but not re-inspecting it on re-entry,
and also occasionally reading it not under lock.  This CL makes the locking
story more uniform and eliminates known bugs.

BUG=144319


Review URL: https://chromiumcodereview.appspot.com/10876049