Revert 153206 - Re-enable Ctrl-w and Ctrl-Shift-w when a non-Pepper window is maximized.
This is a temporary fix for M22 for crbug.com/143442, and should be reverted on trunk as part of crbug.com/142422. This CL is basically a second trial of r151121 which was also a temporary fix for M22.  r151121 disabled Ctrl-w and Ctrl-Shift-w when any kind of full screen window was active, but it turned out that the limitation was too strict; we should not disable Ctrl-w and Ctrl-Shift-w when a non-Pepper window was maximized.

BUG=143442,142422
TEST=confirmed that C-w/C-S-w do not crash when a browser window is maximized.

Review URL: https://chromiumcodereview.appspot.com/10834421

TBR=yusukes@chromium.org
Review URL: https://chromiumcodereview.appspot.com/10867053