New histograms to track disk cache hit rate by age and use time.
Previously we reported the age of the disk cache, and we reported the
hit rate of the disk cache, but there was no statistic that let us
correlate the two. These histograms let us see how the hit rate in a
cache varies with its age; perhaps we'll learn that there's an
equilibrium hit rate, or perhaps we'll learn how long it takes to
reach that.

R=rvargas@chromium.org,tburkard@chromium.org
BUG=None


Review URL: https://chromiumcodereview.appspot.com/10834361