[content shell] improve support of the test shell protocol
This should reduce the amounts of timeouts.

The test shell protocol expects two responses per test, one is a text/html or
audio response, and the other is a image response (potentially without an
actual image). Each block is terminated by #EOF. In addition, an #EOF has to be
send to stderr as well.

Also clean up responsibilities a bit, so WebKitTestController does all the
controlling and WebKitTestRunnerHost merely forwards invocations of the test
runner.

BUG=111316
TEST=manually


Review URL: https://chromiumcodereview.appspot.com/10894010