Add target for webkit_compositor_unittests
We need a target for unit tests that test the compositor WebKit APIs but have
knowledge of the implementation (CC guts). This sets up the test target and
puts a few tests in place. Like cc_unittests, this target doesn't work in the
component build and won't until we break link-time dependencies on non-exported
WebCore symbols.

BUG=

Review URL: https://chromiumcodereview.appspot.com/10874095