Fix a timing assumption that is broken by decoupling History and Bookmarks.
Previously, the HistoryQuickProvider unit tests were relying on the fact that the InMemoryURLIndex used by the HistoryService is initialized by the time Bookmarks are done loading. Once that assumption is invalidated, the tests would populate the index only to have it cleared once the HistoryService initialization completed.

This CL adds a method to TestingProfile that allows one to block until the index is populated.

R=sky
BUG=144050

Review URL: https://chromiumcodereview.appspot.com/10887019