Revert 154134 - Change NaCl IPC PPAPI proxy startup to support a NaCl-Browser process channel.
This fails to build on cros_daisy because of OVERRIDE.

NaClProcessHost now creates an initial NaCl-Browser channel, then uses
it to send a message to create the NaCl-Renderer channel. The main() for
the IPC-IRT creates a PpapiDispatcher object to manage this channel and
manage the PluginDispatchers for each renderer.
BUG=116317
TEST=manual
Review URL: https://chromiumcodereview.appspot.com/10883047

TBR=bbudge@chromium.org
Review URL: https://chromiumcodereview.appspot.com/10913007