Make GpuDataManager thread safe.
Instead of forcing it on UI thread only, this CL makes it thread safe.

We already put a locker around gpu_info, this CL puts another locker around log_messages, and the updater_list is thread safe.  Therefore, the whole class now is thread safe.

Also, I did a little bit clean up, 1) remove two dead function declarations; 2) rename Features to BlacklistedFeatures, because the old naming is confusing and multiple people complained to me already. 3) Expose AddLogMessage() so it can also be called on browser thread, which is the original motivation for me to do this refactoring.

I am working on a unit test for GpuDataManager, but will upload it as a separate CL.  This CL does not change behavior at all.

BUG=
TEST=tree
Review URL: https://chromiumcodereview.appspot.com/10891013