ash: Add some gesture support to the shelf.
With this change, the shelf can be hidden by dragging it down (in horizontal
position, left/right when in vertical position). The ShelfGestureHandler is
currently used by the launcher and tray widgets. In a follow-up CL, the
bezel-gesture handler will also use it to control the shelf.

BUG=145406

Review URL: https://chromiumcodereview.appspot.com/10916004